Moved typing to opensearch package
