+++ sciety_labs/app/routers/api/categorisation/providers.py
@@ -11,7 +11,7 @@ from sciety_labs.app.routers.api.categorisation.typing import (
     CategorisationResponseDict
 )
 from sciety_labs.providers.opensearch_article_recommendation import DocumentDict
-from sciety_labs.providers.opensearch_typing import OpenSearchSearchResultDict
+from sciety_labs.providers.opensearch.opensearch_typing import OpenSearchSearchResultDict
 
 
 LOGGER = logging.getLogger(__name__)
