+++ sciety_labs/providers/opensearch/opensearch_typing.py
